Fix BUILD.bazel to keep generated headers private
This adds support for the jit extensions as well as making sure that the
generated files do not propagate through the dependency chain.

For example we do not want dependencies (glib-2.0) pick up the config
file from PCRE2

Contains a series of fixes to build on windows, as it seems that bazel
does not like to copy files and refer to them.

Bug: 310438041
Change-Id: I6bd63b40269ca01a966a48e2fc43bc3cd5b6aab4